+++ src/api/Httpd.cpp
@@ -31,8 +31,9 @@
 #include "log/Log.h"
 
 
-Httpd::Httpd(int port, const char *accessToken) :
+Httpd::Httpd(int port, const char *accessToken, bool IPv6) :
     m_idle(true),
+    m_IPv6(IPv6),
     m_accessToken(accessToken ? strdup(accessToken) : nullptr),
     m_port(port),
     m_daemon(nullptr)
@@ -61,6 +62,10 @@ bool Httpd::start()
     }
 
     unsigned int flags = 0;
+    if (m_IPv6 && MHD_is_feature_supported(MHD_FEATURE_IPv6)) {
+        flags |= MHD_USE_DUAL_STACK;
+    }
+
     if (MHD_is_feature_supported(MHD_FEATURE_EPOLL)) {
         flags |= MHD_USE_EPOLL_LINUX_ONLY;
     }